requirements.txt
Django==4.2.9
psycopg2-binary
python-decouple
python-dotenv
Pillow==10.1.0

# Code quality
# ------------------------------------------------------------------------------
black==23.11.0
django-model-utils==4.3.1
flake8==6.1.0
flake8-isort==6.1.1
pytest==7.4.3
pytest-django==4.7.0
pytest-sugar==0.9.7
factory-boy==3.2.1
coverage==7.3.2
django-coverage-plugin==3.1.0 
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
pylint-django==2.5.5
pylint-celery==0.3
mypy==1.7.1
django-allauth==0.61.1
